pyproject.toml
[tool.poetry]
name = "lofin365-data-collector"
version = "1.0.0"
description = "Korean Local Finance Data Collector for expenditure data by sub-projects (2016-2024)"
authors = ["2kwonhee <id@senterie.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/idencosmos/lofin365-data-collector"
keywords = ["finance", "data-collection", "korea", "local-government"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
